Implement share logs in pastila.nl with end-to-end encryption
